Arquitectura modular en desktop_hybrid, Mejor flujo en login
