+++ desktop_hybrid/app/network.py
@@ -1,0 +1,12 @@
+import socket
+from contextlib import closing
+
+
+def is_port_available(host: str, port: int) -> bool:
+    with closing(socket.socket(socket.AF_INET, socket.SOCK_STREAM)) as sock:
+        sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+        try:
+            sock.bind((host, port))
+            return True
+        except OSError:
+            return False
