+++ desktop_hybrid/static/css/style.css
@@ -193,6 +193,7 @@ body {
 }
 .view { display: flex; flex-direction: column; gap: 0; }
 .view h2 { font-size: 1.15rem; font-weight: 900; color: var(--text); margin-bottom: 4px; }
+.view.view-camera { gap: 0; }
 
 /* ── Stats home ── */
 .grid-2 { display: grid; grid-template-columns: 1fr 1fr; gap: 10px; }
@@ -275,6 +276,21 @@ body {
 }
 video { width: 100%; height: 100%; display: block; object-fit: cover; }
 
+.camera-stage {
+  position: relative;
+  margin: -14px -12px 0;
+  height: calc(100vh - var(--topbar-h));
+}
+
+.camera-box.camera-full {
+  height: 100%;
+  margin: 0;
+  border-radius: 0;
+  border: none;
+  box-shadow: none;
+  aspect-ratio: auto;
+}
+
 .camera-overlay {
   position: absolute; inset: 0;
   background: rgba(0,0,0,.6);
@@ -288,6 +304,47 @@ video { width: 100%; height: 100%; display: block; object-fit: cover; }
 }
 .camera-overlay__txt { font-size: .78rem; color: rgba(255,255,255,.5); font-weight: 600; }
 
+.camera-hud {
+  position: absolute; inset: 0;
+  display: flex; flex-direction: column; justify-content: space-between;
+  padding: 16px;
+  background: linear-gradient(180deg, rgba(0,0,0,.55) 0%, rgba(0,0,0,.15) 38%, rgba(0,0,0,.55) 100%);
+  pointer-events: none;
+}
+
+.camera-hud__top,
+.camera-hud__bottom {
+  display: flex; flex-direction: column; gap: 10px;
+}
+
+.camera-hud h2,
+.camera-sub,
+.camera-hud .angle-hint {
+  color: #fff;
+}
+
+.camera-sub {
+  font-size: .82rem;
+  font-weight: 600;
+  color: rgba(255,255,255,.82);
+}
+
+.camera-pill {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  width: fit-content;
+  padding: 6px 12px;
+  border-radius: 999px;
+  background: rgba(255,255,255,.16);
+  color: #fff;
+  font-size: .64rem;
+  font-weight: 800;
+  letter-spacing: .08em;
+  text-transform: uppercase;
+  border: 1px solid rgba(255,255,255,.35);
+}
+
 /* Liveness hint */
 .liveness-hint {
   display: flex; align-items: flex-start; gap: 10px;
@@ -300,8 +357,21 @@ video { width: 100%; height: 100%; display: block; object-fit: cover; }
 }
 .liveness-hint__txt { font-size: .78rem; font-weight: 700; color: var(--text-mid); line-height: 1.35; }
 
+.camera-hud .liveness-hint {
+  margin: 0;
+  background: rgba(0,0,0,.45);
+  border: 1px solid rgba(255,255,255,.25);
+}
+
+.camera-hud .liveness-hint__txt { color: rgba(255,255,255,.85); }
+
 /* ════ BOTONES ════ */
 .row { display: flex; gap: 8px; flex-wrap: wrap; margin: 8px 0; }
+
+.camera-actions {
+  margin: 0;
+  pointer-events: auto;
+}
 
 button, .btn {
   display: inline-flex; align-items: center; gap: 7px;
@@ -317,7 +387,7 @@ button:active   { transform: scale(.97); }
 button:disabled { opacity: .38; cursor: not-allowed; transform: none; box-shadow: none; background: var(--primary); }
 
 .btn-full   { width: 100%; justify-content: center; padding: 14px; font-size: .95rem; }
-.btn-danger { background: #B91C1C; box-shadow: 0 4px 12px rgba(185,28,28,.26); }
+.btn-danger {  width: 100%; justify-content: center; background: #B91C1C; box-shadow: 0 4px 12px rgba(185,28,28,.26); }
 .btn-danger:hover { background: #991B1B; box-shadow: 0 6px 16px rgba(185,28,28,.36); }
 .btn-back {
   background: transparent; color: var(--muted);
@@ -337,6 +407,83 @@ button:disabled { opacity: .38; cursor: not-allowed; transform: none; box-shadow
 .feedback.granted { background: var(--ok-bg); border-color: var(--ok-border); color: var(--ok-text); }
 .feedback.denied  { background: var(--err-bg); border-color: var(--err-border); color: var(--err-text); }
 .feedback.waiting { background: #F4FBF6; border-color: var(--line); color: var(--muted); }
+
+.camera-hud .feedback {
+  margin: 0;
+  color: #fff;
+  background: rgba(0,0,0,.55);
+  border: 1px solid rgba(255,255,255,.25);
+}
+
+.camera-hud .feedback.granted {
+  background: rgba(6,78,59,.7);
+  border-color: rgba(110,231,183,.6);
+}
+
+.camera-hud .feedback.denied {
+  background: rgba(127,29,29,.75);
+  border-color: rgba(252,165,165,.6);
+}
+
+.camera-hud .feedback-help {
+  font-size: .78rem;
+  color: rgba(255,255,255,.85);
+  background: rgba(0,0,0,.45);
+  border: 1px solid rgba(255,255,255,.2);
+  border-radius: 10px;
+  padding: 8px 12px;
+}
+
+.camera-hud .feedback-help.is-clickable {
+  text-decoration: underline;
+  cursor: pointer;
+  pointer-events: auto;
+}
+
+.help-modal.hidden { display: none; }
+.help-modal {
+  position: fixed; inset: 0;
+  z-index: 500;
+  display: flex; align-items: center; justify-content: center;
+  padding: 20px;
+}
+.help-modal__overlay {
+  position: absolute; inset: 0;
+  background: rgba(4,14,8,.7);
+  backdrop-filter: blur(2px);
+}
+.help-modal__card {
+  position: relative;
+  width: min(420px, 92vw);
+  background: #fff;
+  border-radius: 16px;
+  border: 2px solid var(--line);
+  box-shadow: 0 12px 36px rgba(0,60,20,.25);
+  padding: 16px;
+  z-index: 1;
+}
+.help-modal__head {
+  display: flex; align-items: center; justify-content: space-between;
+  gap: 12px;
+  margin-bottom: 12px;
+}
+.help-modal__head h3 {
+  font-size: 1rem;
+  font-weight: 900;
+  color: var(--text);
+}
+.help-modal__body ul {
+  margin: 0 0 10px 18px;
+  color: var(--text-mid);
+  font-size: .86rem;
+  font-weight: 600;
+  line-height: 1.5;
+}
+.help-modal__note {
+  font-size: .78rem;
+  color: var(--muted);
+  font-weight: 700;
+}
 
 /* ════ ACCESO: resultado ════ */
 .access-result { display: flex; flex-direction: column; gap: 12px; }
